250430 Move request DTOs to inner classes
