--- src/main/java/backend/time/dto/ActivityNotificationDto.java
@@ -1,16 +1,0 @@
-package backend.time.dto;
-
-import backend.time.model.ActivityNotification;
-import backend.time.model.ActivityType;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-public class ActivityNotificationDto {
-
-    List<ActivityNotificationListDto> activityNotificationListDtoList;
-
-}
