--- src/main/java/backend/time/dto/BoardDistanceDto.java
@@ -1,12 +1,0 @@
-package backend.time.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class BoardDistanceDto {
-    private Long id;
-    private Double distance;
-}
-
