--- src/main/java/backend/time/dto/KakaoDto.java
@@ -1,14 +1,0 @@
-package backend.time.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class KakaoDto {
-    String kakaoId;
-}
