--- src/main/java/backend/time/dto/KeywordNotificationDto.java
@@ -1,12 +1,0 @@
-package backend.time.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class KeywordNotificationDto {
-
-    private List<KeywordNotificationListDto> keywordNotificationListDtos;
-
-}
