--- src/main/java/backend/time/dto/RoomEnterDto.java
@@ -1,10 +1,0 @@
-package backend.time.dto;
-
-import lombok.Data;
-
-@Data
-public class RoomEnterDto {
-
-    private String roomName;
-    private Long boardId;
-}
