--- src/main/java/backend/time/dto/UnfinishedMemberDto.java
@@ -1,15 +1,0 @@
-package backend.time.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UnfinishedMemberDto {
-    String kakaoId;
-    String nickname;
-}
