+++ src/main/java/backend/time/dto/request/ChatRequestDto.java
@@ -1,0 +1,12 @@
+package backend.time.dto.request;
+
+import lombok.Data;
+
+public class ChatRequestDto {
+    @Data
+    public static class RoomEnterDto {
+
+        private String roomName;
+        private Long boardId;
+    }
+}
