--- src/main/java/backend/time/dto/request/PayMethDto.java
@@ -1,12 +1,0 @@
-package backend.time.dto.request;
-
-import lombok.Data;
-
-@Data
-public class PayMethDto {
-    private String payMeth;
-    //계좌이체일때
-    private String holder; // 예금주
-    private String bank; // 은행
-    private Long accountNumber; //계좌번호
-}
