--- src/main/java/backend/time/dto/request/PointDto.java
@@ -1,12 +1,0 @@
-package backend.time.dto.request;
-
-import lombok.Data;
-
-@Data
-public class PointDto {
-
-    private Double longitude;
-    private Double latitude;
-    private String address;
-    private String kakaoId;
-}
