--- src/main/java/backend/time/dto/request/PostPrepareDto.java
@@ -1,9 +1,0 @@
-package backend.time.dto.request;
-
-import lombok.Data;
-
-@Data
-public class PostPrepareDto {
-    private String merchant_uid;
-    private Long amount;
-}
