+++ src/main/java/backend/time/dto/request/ReportRequestDto.java
@@ -1,8 +1,6 @@
 package backend.time.dto.request;
 
 import backend.time.model.ReportCategory;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ReportDto {
+public class ReportRequestDto {
     private ReportCategory reportCategory; // 신고 사유
 
 }
