+++ src/main/java/backend/time/dto/request/ScrapRequestDto.java
@@ -3,6 +3,6 @@ package backend.time.dto.request;
 import lombok.Data;
 
 @Data
-public class ScrapDto {
+public class ScrapRequestDto {
     private int pageNum;
 }
