--- src/main/java/backend/time/dto/request/VerifyAndChargeDto.java
@@ -1,8 +1,0 @@
-package backend.time.dto.request;
-
-import lombok.Data;
-
-@Data
-public class VerifyAndChargeDto {
-    private String merchant_uid;
-}
