+++ src/main/java/backend/time/repository/CustomBoardRepositoryImpl.java
@@ -3,7 +3,7 @@ package backend.time.repository;
 import static backend.time.model.board.QBoard.board;
 import static backend.time.model.board.QImage.image;
 
-import backend.time.dto.request.BoardSearchDto;
+import backend.time.dto.request.BoardRequestDto.BoardSearchDto;
 import backend.time.dto.response.BoardResponseDto.BoardSearchHaversine;
 import backend.time.dto.response.BoardResponseDto.BoardSearchSpatial;
 import backend.time.dto.response.QBoardResponseDto_BoardSearchHaversine;
