+++ src/main/java/backend/time/service/NotificationService.java
@@ -3,6 +3,10 @@ package backend.time.service;
 import backend.time.controller.MyPageController;
 import backend.time.controller.NotificationController;
 import backend.time.dto.*;
+import backend.time.dto.response.NotificationResponseDto.ActivityNotificationDto;
+import backend.time.dto.response.NotificationResponseDto.ActivityNotificationListDto;
+import backend.time.dto.response.NotificationResponseDto.KeywordNotificationDto;
+import backend.time.dto.response.NotificationResponseDto.KeywordNotificationListDto;
 import backend.time.model.*;
 import backend.time.model.Member.Member;
 import backend.time.model.board.Board;
